[backend] new file filtering-completeSpecialFilterKeys
